move _configure_legends to structure_viz/helpers.py
